minor refactoring, renamed GardenList to Gardens
