+++ client/App.jsx
@@ -10,9 +10,9 @@ import AddEvent from './pages/admin/AddEvent/AddEvent'
 import EditEvent from './pages/admin/EditEvent/EditEvent'
 import Error from './components/Error/Error'
 import Event from './pages/Event/Event'
+import Gardens from './pages/Gardens/Gardens'
 
 import { cacheUser } from './auth-utils'
-import GardenList from './components/GardenList/GardenList'
 
 export default function App () {
   cacheUser(useAuth0)
@@ -23,7 +23,7 @@ export default function App () {
       <Header />
       <main className='container'>
         <Route exact path='/' component={Home} />
-        <Route exact path='/gardens' component={GardenList} />
+        <Route exact path='/gardens' component={Gardens} />
         <Route exact path='/gardens/:id' component={Garden} />
         <Route exact path='/gardens/:id/events/:eventId' component={Event} />
         <Route path='/profile' component={Profile} />
