+++ client/pages/Gardens/Gardens.test.js
@@ -2,10 +2,10 @@ import React from 'react'
 import { screen } from '@testing-library/react'
 import { renderWithRouter } from '../../test-utils'
 
-import GardenList from './GardenList'
-import { getAllGardens } from './gardernListHelper'
+import Gardens from './Gardens'
+import { getAllGardens } from './gardensHelper'
 
-jest.mock('./gardernListHelper')
+jest.mock('./gardensHelper')
 
 afterEach(() => {
   getAllGardens.mockClear()
@@ -17,7 +17,7 @@ describe('list of gardens', () => {
       return Promise.resolve([])
     })
 
-    renderWithRouter(<GardenList />)
+    renderWithRouter(<Gardens />)
 
     const list = await screen.findByRole('list')
     expect(list).toBeEmptyDOMElement()
@@ -38,7 +38,7 @@ describe('list of gardens', () => {
       ])
     })
 
-    renderWithRouter(<GardenList />)
+    renderWithRouter(<Gardens />)
 
     return screen.findAllByRole('listitem')
       .then((listItems) => {
@@ -75,7 +75,7 @@ describe('list of gardens', () => {
       ])
     })
 
-    renderWithRouter(<GardenList />)
+    renderWithRouter(<Gardens />)
 
     return screen.findAllByRole('listitem')
       .then((listItems) => {
