+++ client/pages/Gardens/gardensHelper.test.js
@@ -1,5 +1,5 @@
 import { dispatch } from '../../store'
-import { getAllGardens } from './gardernListHelper'
+import { getAllGardens } from './gardensHelper'
 import { SET_WAITING, CLEAR_WAITING } from '../../actions/waiting'
 import { SHOW_ERROR } from '../../actions/error'
